In remarkup, strip one leading space from each line in a quoted block if possible
Summary:
Fixes T13268. Currently, when we quote text, we add `> ` (with a space) but don't try to strip it when un-quoting text.

Instead, try to strip it.

We can still end up with some occasional weird cases (users manually typing ">") that we may have to refine eventually, but this should make the common case ("Quote" button in the UI) work more consistently.

Test Plan:
Added a test, made it pass.

{F6312077}

Reviewers: amckinley

Reviewed By: amckinley

Maniphest Tasks: T13268

Differential Revision: https://secure.phabricator.com/D20327