Reprocess Feeders Post Rate Type Addition
Summary:
Currently, when a new rate type is added, we need to reprocess the feeders to feed the data for the newly added rate type. Hence created a process to reprocess the feeders upon addition of a new rate type. If a new rate type is added, it
will be updated in admin playbook. Based on that condition, this process will be triggered in the batch 1025

Test Plan: https://docs.google.com/document/d/1CVp9bwsIPO9K2jzLyWEm9AB-LsmNJK3bOnd0zNRa6f0/edit

Reviewers: #tm1_application_platform_tap, #ldap_tm1, balajik

Reviewed By: #tm1_application_platform_tap, #ldap_tm1, balajik

Subscribers: #ldap_tm1

JIRA Issues: FINTELPA-5727

Differential Revision: https://code.uberinternal.com/D8615615