Update ART to remove Cash Trips
Summary:
A Separate ART is already present to recon the cash trips between ops and analytics. This ART has cashtrips in it but it doesnt liase with the sync timings of cash trips. Hence causing tming issues during recon on BD+1 and BD+2
So updating the ART to remove the cash trips from it

Test Plan: Attached in JIRA

Reviewers: #tm1_application_platform_tap, #ldap_tm1, balajik

Reviewed By: #tm1_application_platform_tap, #ldap_tm1, balajik

Subscribers: #ldap_tm1

JIRA Issues: FINTELPA-6263

Differential Revision: https://code.uberinternal.com/D9015631