Trips Data load process update
Summary: Currently we are copying both Week and Month level data to Ops cube from Mobility/Delivery cubes. As part of the new requirements we need to copy day level data also to Ops cube. In this diff we will be calling two separate processes to copy Month data in one process and Day/Week in separate process so that once Month is copied we can directly run the analytics sync instead of waiting for Day/Week sync.

Reviewers: #tm1_application_platform_tap, #ldap_tm1, gurudamodharan

Reviewed By: #tm1_application_platform_tap, #ldap_tm1, gurudamodharan

Subscribers: #ldap_tm1

JIRA Issues: FINTELPA-6718

Differential Revision: https://code.uberinternal.com/D9385953