PAX redundant objects build update
Summary:
As part of PAX redundant objects, we are calling the redundant objects handling TI at the end of the build. The current code has a bug that it is calling the TI in the replica instance instead of the Main instance. Hence updating the code
to call this in Main instance

Test Plan: https://docs.google.com/spreadsheets/d/1WhsgiD4gNnu3fxhZWOcSBgnuPOl7Sx87cKeJvd0_zoM/edit#gid=366729608

Reviewers: #tm1_application_platform_tap, #ldap_tm1, ajitsingh7

Reviewed By: #tm1_application_platform_tap, #ldap_tm1, ajitsingh7

Subscribers: #ldap_tm1

JIRA Issues: FINTELPA-7022

Differential Revision: https://code.uberinternal.com/D9650103