Mobility Forecasting allocations PAW subsets Update
Summary: As part of Mobility Forecast allocations, we are maintaining a PAW workbook . This PAW workbook has various reports from various cubes that are related to allocations. we are creating static subsets that needs to be updated constantly based on certain factors. Eg., if a new country is added, the country susbet needs to be update., If we move to the net month, the forecast months and the current forecast month needs to be updated. Hence this process takes care of creating the subsets dynamically

Test Plan: https://paw.tm1uat.awscorp.uberinternal.com/?perspective=dashboard&id=7757c6e4-7e5e-40f4-bc91-15fc56459e68

Reviewers: #tm1_application_platform_tap, #ldap_tm1, balajik

Reviewed By: #tm1_application_platform_tap, #ldap_tm1, balajik

Subscribers: #ldap_tm1

JIRA Issues: FINTELPA-7165

Differential Revision: https://code.uberinternal.com/D9962591