Mobility and Delivery - product pl Actual allocations
Summary: As part of Product PL Allocations, we allocate the oracle data down to product type and features. The reference for this allocation process was sourced from "analytics.gl operational" cube. This was the design pre NETR as the dimensionality of analytics and ops were different. Now that the dimensions are in sync, we are updating the TI to source from the mobility cube itself. This way we can clean up the ananlytics.gl ops cube in prod and that would save great amount of space and memory

Test Plan: https://docs.google.com/spreadsheets/d/1uItH2bTrypNdyVc8zsbwl9xGRuoD6dsAEwioiVS3cfw/edit#gid=1644270586

Reviewers: #tm1_application_platform_tap, #ldap_tm1, ajitsingh7

Reviewed By: #tm1_application_platform_tap, #ldap_tm1, ajitsingh7

Subscribers: #ldap_tm1

JIRA Issues: FINTELPA-5938

Differential Revision: https://code.uberinternal.com/D11178667