MAPC - Alternate hierarchy Cleanup
Summary: As Part of MAPC implementation, we cannot aggregate the data naturally in TM1 consolidation. So we place the data in virtual hierarchy and use TM1 Rules to populate them. But one of the limitation in TM1 is that we cannot use the generic TI Process to zero out the data in the alternate hierarchy. Hence creating this separate TI to zero out the data in the alternate hierarchy

Test Plan: https://docs.google.com/spreadsheets/d/1FXzBWuMx3l8CXCDwlPD5AarNMZHVnwHNZyetLDoZcYg/edit#gid=0

Reviewers: #tm1_application_platform_tap, #ldap_tm1, ramakrishnan

Reviewed By: #tm1_application_platform_tap, #ldap_tm1, ramakrishnan

Subscribers: #ldap_tm1

JIRA Issues: FINTELPA-8486

Differential Revision: https://code.uberinternal.com/D11796273