mapc forecast clean up at alternate hierarhcy - TI
Summary:
As part of MAPC Phase 3 project, this TI process is to zero out forecast MAPC metrics (Active Users & New Users) at the alternate hierarchy of LOB and Location dimensions.
This TI puts 0 for the MAPC metrics at the alternate hierarchy intersection of LOB and Location dimensions. This is needed to be done before we import the deduped forecast numbers from Big Data.

Test Plan:
attaching the main view from testing performed in UAT as cannot show the data clearing part. It occurs when we load the data as a pre-step.

{F896664819}

{F896664891}

Reviewers: #tm1_application_platform_tap, #ldap_tm1, vikas.dwivedi

Reviewed By: #tm1_application_platform_tap, #ldap_tm1, vikas.dwivedi

Subscribers: #ldap_tm1

JIRA Issues: FINTELPA-9156

Differential Revision: https://code.uberinternal.com/D13455845