copy forecast MAPCs from old to new mapc metrics
Summary: as part of mapc phase 3 project, we need to copy the MAPC forecast data from Active Riders and New Riders to Active Users and New Users respectively. This TI does the job taking in year or month as the parameter.

Test Plan:
validation perfomed in UAT is attached.

{F896666797}

Reviewers: #tm1_application_platform_tap, #ldap_tm1, vikas.dwivedi

Reviewed By: #tm1_application_platform_tap, #ldap_tm1, vikas.dwivedi

Subscribers: #ldap_tm1

JIRA Issues: FINTELPA-9156

Differential Revision: https://code.uberinternal.com/D13455883